+++ src/app/sensors.cpp
@@ -12,6 +12,10 @@
 #include "sensors/ld2410_out_sensor.cpp"
 #endif
 
+#if HAS_SENSOR_DUMMY
+#include "sensors/dummy_sensor.cpp"
+#endif
+
 void sensor_manager_register_all(SensorRegistry &registry) {
 	#if HAS_SENSOR_BME280
 	register_bme280_sensor(registry);
@@ -19,6 +23,10 @@ void sensor_manager_register_all(SensorRegistry &registry) {
 
 	#if HAS_SENSOR_LD2410_OUT
 	register_ld2410_out_sensor(registry);
+	#endif
+
+	#if HAS_SENSOR_DUMMY
+	register_dummy_sensor(registry);
 	#endif
 }